+++ da-indexer/da-indexer-logic/src/celestia/da.rs
@@ -1,5 +1,5 @@
 use crate::{
-    celestia::{repository::blobs, rpc_client, rpc_client::ShareV2Client},
+    celestia::{client, repository::blobs},
     indexer::{Job, DA},
 };
 use anyhow::Result;
@@ -12,7 +12,10 @@ use std::sync::{
     Arc,
 };
 
-use super::{job::CelestiaJob, parser, repository::blocks, settings::IndexerSettings};
+use super::{
+    client::share::ShareClient, job::CelestiaJob, parser, repository::blocks,
+    settings::IndexerSettings,
+};
 
 pub struct CelestiaDA {
     client: Client,
@@ -24,7 +27,7 @@ pub struct CelestiaDA {
 
 impl CelestiaDA {
     pub async fn new(db: Arc<DatabaseConnection>, settings: IndexerSettings) -> Result<Self> {
-        let client = rpc_client::new_celestia_client(
+        let client = client::new_celestia_client(
             &settings.rpc.url,
             settings.rpc.auth_token.as_deref(),
             settings.rpc.max_request_size,
@@ -52,14 +55,13 @@ impl CelestiaDA {
     }
 
     async fn get_blobs_by_height(&self, height: u64) -> Result<(ExtendedHeader, Vec<Blob>)> {
-        // TODO: it seems possible to avoid this request with new Celestia API
         let header = self.client.header_get_by_height(height).await?;
 
         let mut blobs = vec![];
         if parser::maybe_contains_blobs(&header.dah) {
             let eds = self
                 .client
-                .share_get_eds_v2(height, header.header.version.app)
+                .share_get_eds(height, header.header.version.app)
                 .await?;
             blobs = parser::parse_eds(&eds, header.header.version.app)?;
         }
@@ -105,6 +107,11 @@ impl DA for CelestiaDA {
     async fn new_jobs(&self) -> anyhow::Result<Vec<Job>> {
         let height = self.client.header_local_head().await?.header.height.value();
         tracing::info!(height, "latest block");
+
+        if height <= self.last_known_height.load(Ordering::Acquire) {
+            tracing::info!("latest block is below last known height, skipping...");
+            return Ok(vec![]);
+        }
 
         let from = self.last_known_height.swap(height, Ordering::AcqRel) + 1;
         Ok((from..=height)
