+++ da-indexer/da-indexer-logic/src/celestia/mod.rs
@@ -1,9 +1,9 @@
+mod client;
 pub mod da;
 pub mod job;
 pub mod l2_router;
 mod parser;
 pub mod repository;
-mod rpc_client;
 pub mod settings;
 #[cfg(test)]
 pub mod tests;
